+++ src/bp/core/user-code/action-service.ts
@@ -3,7 +3,7 @@ import { IO, Logger } from 'botpress/sdk'
 import { extractEventCommonArgs } from 'common/action'
 import { ObjectCache } from 'common/object-cache'
 import { ActionScope, ActionServer, LocalActionDefinition } from 'common/typings'
-import { createForAction } from 'core/api'
+import { createForAction } from 'core/app/api'
 import { BotService } from 'core/bots'
 import { GhostService } from 'core/bpfs'
 import { ActionExecutionError } from 'core/dialog'
